Move input validators to validator module for artifacts
